Final commit for the night. - Bug is now a viable enemy - Battle map only loads the enemy that you interacted with. Other NPC are present for debug - Implemented health and strength in enemy class - Health label should change with damage received - Game over when player health dips to zero & below

+++ src/NPCs/Shrek.java
@@ -20,7 +20,7 @@ public class Shrek extends Enemy {
         super(id, location.x, location.y, new SpriteSheet(ImageLoader.load("Shrek.png"), 40, 40), "STAND_RIGHT", 10, 1);
     }
 
-    public Shrek(int id, Point location, int health, int strength) {
+    public Shrek(int id, Point location, float health, float strength) {
         super(id, location.x, location.y, new SpriteSheet(ImageLoader.load("Shrek.png"), 40, 40), "STAND_RIGHT", health, strength);
     }
